[f] 251 - Support float value for visual token color ranges
[#172258251](https://www.pivotaltracker.com/story/show/172258251)

+++ cpdb/app_config/models/visual_token_color.py
@@ -8,8 +8,8 @@ from app_config.validators import is_valid_hex_color
 class VisualTokenColor(TimeStampsModel):
     color = models.CharField(max_length=7, validators=[is_valid_hex_color])
     text_color = models.CharField(max_length=7, validators=[is_valid_hex_color])
-    lower_range = models.IntegerField(validators=[MinValueValidator(0), MaxValueValidator(100)])
-    upper_range = models.IntegerField(validators=[MinValueValidator(0), MaxValueValidator(100)])
+    lower_range = models.FloatField(validators=[MinValueValidator(0), MaxValueValidator(100)])
+    upper_range = models.FloatField(validators=[MinValueValidator(0), MaxValueValidator(100)])
 
     def __str__(self):
         return f'[{self.lower_range} - {self.upper_range}]: { self.color}'
